refactor(editor): rename project-roles feature folder to roles with instance/project/shared subfolders
