+++ packages/frontend/editor-ui/src/features/roles/project/ProjectRoleView.vue
