+++ packages/frontend/editor-ui/src/features/roles/project/RoleAssignmentsTab.vue
