+++ packages/frontend/editor-ui/src/features/roles/project/projectRoleScopes.ts
